Sign/requirements.txt
Flask
pymongo
redis
python-dotenv
PyJWT
bcrypt
pydantic[nguyentienthinh3004@gmail.com]